Changed RepoBrowser options format
There was some confusion as to what was actually to be specified to the
RepoBrowser as the information for the original repositories. The
format of the RepoBrowser constructor options was changed so that it
takes an array of objects with information about each repository,
including a title, url of the file where the repo's manifest can be
found, and the uri of the repo.