Cleaned up TextEditorAnnotate and made highlight spans use RDFa
While there's still a lot of cruft which I'm hesitant to remove for
fear of breaking something, TextEditorAnnotate should look a bit
cleaner.
I also removed the old method of stuffing highlight ids into css
classes, since it wasn't clean and seemed to be causing issues with
special characters. The editor now uses RDFa compliant `about`
properties to indicate the uri of the oa:TextQuoteSelector for the
highlight, as well as an oa:exact predicate so they provide that
information for free. The appropriate namespaces are also bound to the
html document for the editor, although the closure editor doesn't
export the entire html document (body only).
The appropriate RDF triples are also generated for the highlights,
although there is not yet a way of keeping the quotes updated (Closes
#24).