Prevented hover menus from failing to show no annotations message when they don't know how to render the anno bodies and targets
This was problematic for things like transcriptions, which technically
target a canvas segment. The AnnoTitlesList class would see that there
were targets, so it wouldn’t display the no annotations message, but
when it went to render the Resource Summaries via the
ResourceSummaryFactory class, nothing would show up. I put a list of
the renderable resource types in the ResourceSummaryFactory module, and
have AnnoTitlesList check against that before adding a resource to its
bodies or target lists